re-implement drag

+++ DragComponent/src/main/java/io/asteroidsfx/dragcomponent/DragComponent.java
@@ -4,6 +4,6 @@ import io.asteroidsfx.common.Component;
 
 public class DragComponent extends Component {
 
-    public double drag;
+    public float drag;
 
 }
